# mock-server/openapi.yaml
openapi: 3.0.0
info:
  title: Smart Clinic OS API Specification (SCS)
  description: |
    المستند البرمجي لواجهات برمجة التطبيقات (API Specification) لمنصة عيادتي الذكية.
    يحتوي الملف على مسارين أساسيين:
    1. **واجهات الأوبريشن والمشغل (Platform Operations):** تبدأ بـ `/admin/v1` ومحمية بـ VPN.
    2. **واجهات العيادات والعملاء (Clinic Dashboard):** تبدأ بـ `/v1` وتعتمد على ترويسة العيادة.
  version: 1.0.0
servers:
  - url: https://api.SCS-admin.com
    description: Public Clinic Dashboard Gateway
  - url: https://api.SCS-ops.com
    description: Private Operations Gateway (Accessible via VPN only)
  - url: http://localhost:3000
    description: Local Mock Development Server
paths:
...
  /v1/sync:
    post:
      summary: مزامنة بيانات الأوفلاين وحل النزاعات (Offline Bulk Sync)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - actions
              properties:
                actions:
                  type: array
                  items:
                    type: object
                    properties:
                      action_id:
                        type: string
                      action_type:
                        type: string
                      timestamp:
                        type: string
                      payload:
                        type: object
      responses:
        '200':
          description: تمت المزامنة بنجاح وبدون تعارضات
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      synced_count:
                        type: integer
                        example: 1
                      conflicts:
                        type: array
                        example: []
        '409':
          description: فشل المزامنة لوجود تعارض حجز مزدوج
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: OFFLINE_SYNC_CONFLICT
                      message:
                        type: string
                        example: يوجد تعارض في البيانات التي تم تعديلها أثناء عدم الاتصال
  /v1/queue/call-next:
    post:
      summary: نداء المريض وتحديث الشاشة (Call Next Patient)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - doctor_id
              properties:
                doctor_id:
                  type: string
                  example: doctor-uuid-12345
      responses:
        '200':
          description: تم النداء وتحديث الشاشة عبر WebSockets
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      called_patient:
                        type: object
                        properties:
                          queue_number:
                            type: integer
                            example: 14
                          display_name:
                            type: string
                            example: محمد أ.
                          appointment_id:
                            type: string
                            example: appointment-uuid-xxxxx
                      websocket_broadcast_sent:
                        type: boolean
                        example: true
  /v1/settings/notifications:
    get:
      summary: الحصول على إعدادات تفضيلات الإشعارات للعيادة (Get Notification Settings)
      responses:
        '200':
          description: تفضيلات الإشعارات الحالية للعيادة
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      notification_settings:
                        type: object
                        properties:
                          patient_email_booking_confirm:
                            type: boolean
                            example: true
                          patient_whatsapp_booking_confirm:
                            type: boolean
                            example: true
                          patient_email_prescription:
                            type: boolean
                            example: true
                          patient_email_invoice:
                            type: boolean
                            example: true
                          doctor_email_new_booking:
                            type: boolean
                            example: true
                          doctor_whatsapp_new_booking:
                            type: boolean
                            example: false
                          doctor_email_daily_report:
                            type: boolean
                            example: true
                          doctor_email_weekly_report:
                            type: boolean
                            example: true
    put:
      summary: تحديث إعدادات تفضيلات الإشعارات للعيادة (Update Notification Settings)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - notification_settings
              properties:
                notification_settings:
                  type: object
      responses:
        '200':
          description: تم التعديل بنجاح
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: تم تحديث إعدادات الإشعارات بنجاح
                      updated_settings:
                        type: object

  # --- Webhook Routes ---
...